fix(projects): keep node:fs out of the client bundle (restore types split)
f811f20 intended a client/server split so "use client" components import
types + pure helpers without dragging node:fs/promises into the browser
bundle — but the cave-projects-types.ts file was lost in a merge, and
projects-view.tsx imported the *value* normalizeProjectRoot straight from
cave-projects.ts (which imports node:fs/promises). Turbopack then failed to
generate the workspace.tsx client chunk ("does not support external modules
(request: node:fs/promises)"), turning main's Frontend build red.

Recreate src/lib/cave-projects-types.ts (CaveProject + normalizeProjectRoot,
no fs), re-export both from cave-projects.ts for existing server importers,
and point projects-view.tsx at the safe module. pnpm build now compiles and
generates all static pages.
